+++ examples/audio.js
@@ -13,6 +13,8 @@ scene("main", () => {
 	// the music might not autoplay cuz some browser won't allow audio start before any user interaction
 	const music = play("OtherworldlyFoe", { loop: true, });
 
+	volume(0.5);
+
 	const label = add([
 		text(),
 	]);
@@ -20,12 +22,17 @@ scene("main", () => {
 	function updateText() {
 		label.text = `
 ${music.paused() ? "paused" : "playing"}
+time: ${music.time().toFixed(2)}
 volume: ${music.volume()}
 detune: ${music.detune()}
 		`.trim();
 	}
 
 	updateText();
+
+	action(() => {
+		updateText();
+	});
 
 	keyPress("space", () => {
 
@@ -35,34 +42,28 @@ detune: ${music.detune()}
 			music.pause();
 		}
 
-		updateText();
 		play("wooosh");
 
 	});
 
 	keyPress("down", () => {
 		music.volume(music.volume() - 0.1);
-		updateText();
 	});
 
 	keyPress("up", () => {
 		music.volume(music.volume() + 0.1);
-		updateText();
 	});
 
 	keyPress("left", () => {
 		music.detune(music.detune() - 100);
-		updateText();
 	});
 
 	keyPress("right", () => {
 		music.detune(music.detune() + 100);
-		updateText();
 	});
 
 	keyPress("escape", () => {
 		music.stop();
-		updateText();
 	});
 
 });